+++ src/FlagshipHooks.tsx
@@ -45,12 +45,6 @@ export const useFsActivate = (
     }, applyEffectScope);
 };
 
-// NOTES:
-/*
-two possible solutions to avoid massive 'activate api' calls:
-1) wrap the 'useFsModificationsCache' in a useEffect and plug correctly the useEffect the way you need
-2) in the JS SDK, make a cache to understand if the activate call already be done before.
-*/
 const safeMode_getCacheModifications = (
     modificationsRequested: FsModifsRequestedList,
     activateAllModifications = false
@@ -123,6 +117,7 @@ export const useFlagship = (options?: UseFlagshipParams): UseFlagshipOutput => {
     } = computedOptions;
     const {
         hasError,
+        setState,
         state: { fsSdk, fsVisitor, status, log }
     } = useContext(FlagshipContext);
     if (hasError) {
@@ -183,7 +178,12 @@ export const useFlagship = (options?: UseFlagshipParams): UseFlagshipOutput => {
 
     const synchronizeModifications = (activate = false): Promise<number> => {
         if (fsVisitor && fsVisitor.synchronizeModifications) {
-            return fsVisitor.synchronizeModifications(activate);
+            return fsVisitor.synchronizeModifications(activate).then((data) => {
+                if (setState) {
+                    setState((s) => ({ ...s, fsModifications: fsVisitor.fetchedModifications }));
+                }
+                return data;
+            });
         }
         logSdkNotReady();
         return new Promise((resolve) => resolve(405));