+++ RELEASENOTES.md
@@ -1,6 +1,16 @@
 # Flagship REACT SDK - Release notes
 
+## ➡️ Version 2.0.7
+
+### Improvements 💪
+
+-   `onUpdate` prop is triggered in a better way and `status` object is giving more data, have a look to the [documentation](https://developers.flagship.io/docs/sdk/react/v2.0#useflagship-output-status) for full details.
+
+### Bug fixes 🐛
+
 ## ➡️ Version 2.0.6
+
+-   `fetchNow` prop was not true by default.
 
 ### Bug fixes 🐛